dist-foreign-service/src/main/java/com/copyright/rup/dist/foreign/service/impl/RightsholderService.java
package com.copyright.rup.dist.foreign.service.impl;

import com.copyright.rup.dist.common.domain.Rightsholder;
import com.copyright.rup.dist.foreign.repository.api.IRightsholderRepository;
import com.copyright.rup.dist.foreign.service.api.IRightsholderService;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.util.List;

/**
 * Implementation of {@link IRightsholderService}.
 * <p/>
 * Copyright (C) 2017 copyright.com
 * <p/>
 * Date: 02/07/2017
 *
 * @author Mikita Hladkikh
 */
@Service
public class RightsholderService implements IRightsholderService {

    @Autowired
    private IRightsholderRepository repository;

    @Override
    public List<Rightsholder> getRros() {
        return repository.findRros();
    }
}
